third_party/WebKit/LayoutTests/batterystatus/basic-operation-expected.txt

CONSOLE MESSAGE: line 10: Uncaught TypeError: Cannot call method 'addEventListener' of undefined
Tests the basic operation of BatteryStatus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
